conda.requirements.txt
gym
git
# gym[atari]
pytorch
scipy
numpy
matplotlib